% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSPoTApi.r
\name{SampleSPoTApi_get_sample_versions}
\alias{SampleSPoTApi_get_sample_versions}
\title{Retrieve a list of object versions by ID}
\usage{
SampleSPoTApi_get_sample_versions(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve a list of object versions by ID
}
